requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pytest
passlib
bcrypt
python-jose
cryptography
httpx